UpcomingMoviesPage moved to specific directory
