+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter, Switch, Route } from 'react-router-dom'
 import * as serviceWorker from './serviceWorker';
-import UpcomingMoviesPage from "./components/UpcomingMoviesPage";
+import UpcomingMoviesPage from "./components/UpcomingMovies/UpcomingMoviesPage";
 import './index.css';
 
 ReactDOM.render(
